auction_web_api/requirements/requirements.txt
djangorestframework==3.15.2
pyJWT==2.9.0
requests==2.32.3